requirements.txt
django
django-ckeditor
django-crispy-forms
django-recaptcha
Pillow